requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
backports.cached-property==1.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.26.164
botocore==1.29.164
Brotli==1.0.9
cachetools==5.3.1
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.4
compressai==1.2.0
cycler==0.11.0
dacite==1.7.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
eventlet==0.33.3
exceptiongroup==1.1.1
fastjsonschema==2.17.1
fastprogress==1.0.3
fiftyone==0.18.0
fiftyone-brain==0.9.2
fiftyone-db==0.4.0
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
gast==0.2.2
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphql-core==3.2.3
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.16.4
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflate64==0.3.1
ipykernel==6.15.0
ipython==7.34.0
ipywidgets==8.1.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter-server-proxy==3.2.2
jupyter-vscode-server==0.0.10
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kaggle==1.6.14
kaleido==0.2.1
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
lightning-utilities==0.9.0
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.1
mongoengine==0.24.2
motor==2.5.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.4
multivolumefile==0.2.3
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
ndjson==0.3.1
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.8.1
numdifftools==0.9.41
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
patool==1.12
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.15.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
py7zr==0.20.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydot==2.0.0
Pygments==2.15.1
pymongo==3.13.0
pyparsing==3.1.0
pyppmd==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
python-slugify==8.0.4
pytorch-lightning==1.9.5
pytorch-msssim==1.0.0
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.1.0
pyzstd==0.15.9
rarfile==4.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
s3transfer==0.6.1
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
semopy==2.3.11
Send2Trash==1.8.2
sentinel==0.3.0
sentry-sdk==2.7.1
setproctitle==1.3.3
simpervisor==0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.20.4
statsmodels==0.13.5
strawberry-graphql==0.138.1
sympy==1.10.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard-pytorch==0.7.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.7
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.8
tinycss2==1.2.1
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
universal-analytics-python3==1.1.1
urllib3==1.26.16
voxel51-eta==0.8.4
wandb==0.17.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
